Move logSetup initialization to build_cmd.go
